refactor(calendar): extract shared helpers to useCalendarState hook
Pull duplicated month-metadata useMemo and day-state helpers (hasWorkout,
isSelected, isToday) out of CalendarWidget and CalendarWidgetMobile into
a shared useCalendarState hook and createDayStateHelpers factory to
prevent desktop-mobile drift.
