+++ src/app/components/CalendarWidgetMobile.tsx
@@ -1,6 +1,9 @@
 import { ChevronLeft, ChevronRight } from "lucide-react";
-import { useMemo } from "react";
 import { cn } from "@/app/components/ui/utils";
+import {
+  createDayStateHelpers,
+  useCalendarState,
+} from "@/app/hooks/useCalendarState";
 
 interface CalendarWidgetMobileProps {
   currentMonth: Date;
@@ -25,18 +28,8 @@ export function CalendarWidgetMobile({
   onDateSelect,
   isDateLocked,
 }: CalendarWidgetMobileProps) {
-  const { daysInMonth, startingDayOfWeek, year, month } = useMemo(() => {
-    const y = currentMonth.getFullYear();
-    const m = currentMonth.getMonth();
-    const firstDay = new Date(y, m, 1);
-    const lastDay = new Date(y, m + 1, 0);
-    return {
-      daysInMonth: lastDay.getDate(),
-      startingDayOfWeek: firstDay.getDay(),
-      year: y,
-      month: m,
-    };
-  }, [currentMonth]);
+  const { daysInMonth, startingDayOfWeek, year, month } =
+    useCalendarState(currentMonth);
 
   const navigateMonth = (direction: "prev" | "next") => {
     // Set day to 1 first to avoid month overflow (e.g., Jan 31 + 1 month = Mar 3)
@@ -49,28 +42,12 @@ export function CalendarWidgetMobile({
     onMonthChange(newDate);
   };
 
-  const isToday = (day: number) => {
-    const today = new Date();
-    return (
-      today.getFullYear() === year &&
-      today.getMonth() === month &&
-      today.getDate() === day
-    );
-  };
-
-  const hasWorkout = (day: number) => {
-    const key = `${year}-${month}-${day}`;
-    return workoutDates.has(key);
-  };
-
-  const isSelected = (day: number) => {
-    if (!selectedDate) return false;
-    return (
-      selectedDate.getFullYear() === year &&
-      selectedDate.getMonth() === month &&
-      selectedDate.getDate() === day
-    );
-  };
+  const { hasWorkout, isSelected, isToday } = createDayStateHelpers(
+    selectedDate,
+    workoutDates,
+    year,
+    month,
+  );
 
   const monthLabel = currentMonth.toLocaleDateString("en-US", {
     month: "long",
